Add files via upload
Initial Submission.  This program defines two functions - calc_average & determine_grade.
The main function controls program flow and calls both the calc_average and
determine_grade functions. The calc_average function accepts 5 test scores as arguments
and returns the average of the scores to the main function. The determine_grade
function accepts a test score as an argument and returns a letter grade for the score
to the main function. The program will display a letter grade for each test score
and the average of the 5 test scores.